Add tests for parsing body of .../complete endpoints, fix decoder (#3)
* Add example json payloads as sent by Chromium

I recorded this by going through the register and login flow, and
dumping the request bodies. We can use this data to write tests for
FromJSON instances, they should be able to handle this.

We'll put them in a directory named after the endpoint, so the test can
check all files in the directory.

* Add tests for parsing body of ./complete endpoints

One of these tests currently fails, the parser contains a bug that I
will fix next.

* Fix bug in authenticator data decoder

The attested credential data is optional, but the decoder assumed it was
always there. Respect the flag that says whether it is present. This
makes the test pass.

* Regenerate fido2.nix