Full Metadata Statement decoding
This implements more-or-less full decoding of Metadata Statements. This
is needed because the IDL description of Metadata Statements represents
e.g. certificates using a text string, which is not fully decoded. If we
want to use those certificates for verification we'd always have to
parse them first. Instead this commit introduces a new Metadata
Statement type and co. that has most fields fully parsed and ready to be
used.

This commit also introduces singletons for the ProtocolFamily, which is
useful for code deduplication and type safety