Pure JWT verification
This commit changes the getPayload function to be jwtToJson instead, and
making that function be pure by allowing the date of verification to be
passed in manually.

However, this date needs to be used by not only the JWT library, but also by
the x509-validation library. Unfortunately the current version of
x509-validation on hackage always gets the current time from IO. So
instead we're using a fork of x509-validation which has support for
doing this purely, the PR for the change is
https://github.com/vincenthz/hs-certificate/pull/126. Notably this pure
function is also needed to have a pure version of attestation
verification, which is going to be done in the future.