Use Validation applicative for asfVerify
There is effectively no need for asfVerify being the Either Monad. By
switching this to Validation we are more explicit about the few early exits,
and it allows us to return multiple validation failures

This opportunity is also used to propagate the attestation statement
validation errors directly, instead of wrapping them in a SomeException.
This allows library users to catch such failures.