Fix AuthenticatorTransport values being ignored
Unknown AuthenticatorTransport values received by the Relying Party need
to still be stored and not cause a decoding error (or be ignored), which
was done previously.

We fix this by introducing a new `AuthenticatorTransportUnknown`
constructor for this case. The `decodeAuthenticatorTransport` is
adjusted to never throw an error since it can handle all inputs now

See also https://github.com/w3c/webauthn/issues/1587